Refactor cache storage/purging
Added the PurgerInterface which is responsible for the actual purging.

No longer need to listen for the CACHE_PURGE event to invalidate a CDN. You were fucked if your command failed because the cache entries are already deleted from the db before the CACHE_PURGE got fired.

The CacheTagsInvalidator no longer immediately purges but rather invalidates the pages by setting it's expiry to the current timestamp. This allows the next cron to purge in batch.

Defining a different storage or purger can be easily done by respectively changing the "wmcontroller.cache.storage" and "wmcontroller.cache.purger" config keys.

See wmcontroller_cloudfront for an example.