Add patch for caching _only_ anonymous users without a CDN.
Currently our implementation expects admins and editors to get
served from a separate CDN where anonymous users don't get cached.
We usually check if the hostname starts with 'admin', and if so
we block access for anonymous users.

We do this because it allows us to shave off a lot of bootstrapping.
Most notably, we don't have to wait for the SessionManager to
do it's thing and create a Session. Waiting for it is especially
stupid when you are 100% certain you'd be serving an anonymous user.

However, without a separate CDN that guarantees we only serve
anonymous users AND without session-info we can't determine
if the user performing the request is anonymous or logged in
nor what role the account has.

So right now you always get served a cached page, even when logged in.

We have a lot of high performance sites and we don't feel comfortable
losing the speed gains we have by not waiting for the session.
Yet we understand not everybody gets to add a separate CDN between
logged-in and non-logged in users. Therefore we added this patch.

Basically it makes sure our Cache middleware runs after the Session middleware.